+++ dkpro-jwpl-api/src/main/java/org/dkpro/jwpl/api/CategoryGraph.java
@@ -40,9 +40,9 @@ import org.dkpro.jwpl.api.exception.WikiApiException;
 import org.dkpro.jwpl.api.exception.WikiPageNotFoundException;
 import org.dkpro.jwpl.api.exception.WikiTitleParsingException;
 import org.dkpro.jwpl.api.util.GraphSerialization;
-import org.dkpro.jwpl.util.ApiUtilities;
-import org.dkpro.jwpl.util.CommonUtilities;
-import org.dkpro.jwpl.util.OS;
+import org.dkpro.jwpl.api.util.ApiUtilities;
+import org.dkpro.jwpl.api.util.CommonUtilities;
+import org.dkpro.jwpl.api.util.OS;
 import org.jgrapht.GraphPath;
 import org.jgrapht.alg.connectivity.ConnectivityInspector;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
