+++ dkpro-jwpl-api/src/main/java/org/dkpro/jwpl/api/Page.java
@@ -25,7 +25,7 @@ import org.dkpro.jwpl.api.exception.WikiPageNotFoundException;
 import org.dkpro.jwpl.api.exception.WikiTitleParsingException;
 import org.dkpro.jwpl.api.hibernate.PageDAO;
 import org.dkpro.jwpl.api.sweble.PlainTextConverter;
-import org.dkpro.jwpl.util.UnmodifiableArraySet;
+import org.dkpro.jwpl.api.util.UnmodifiableArraySet;
 import org.hibernate.LockOptions;
 import org.hibernate.Session;
 import org.hibernate.type.StandardBasicTypes;
