+++ dkpro-jwpl-api/src/main/java/org/dkpro/jwpl/api/WikipediaInfo.java
@@ -28,7 +28,7 @@ import java.util.Set;
 
 import org.dkpro.jwpl.api.exception.WikiApiException;
 import org.dkpro.jwpl.api.exception.WikiPageNotFoundException;
-import org.dkpro.jwpl.util.ApiUtilities;
+import org.dkpro.jwpl.api.util.ApiUtilities;
 import org.hibernate.Session;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
