+++ dkpro-jwpl-api/src/main/java/org/dkpro/jwpl/api/util/DbUtilities.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dkpro.jwpl.util;
+package org.dkpro.jwpl.api.util;
 
 import java.lang.invoke.MethodHandles;
 import java.sql.Connection;
