+++ dkpro-jwpl-api/src/main/java/org/dkpro/jwpl/api/util/distance/LevenshteinStringDistance.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dkpro.jwpl.util.distance;
+package org.dkpro.jwpl.api.util.distance;
 
 public class LevenshteinStringDistance
     implements StringDistance
